Agent Service/requirements.txt
transformers
langchain
torch
scikit-learn
pandas
regex
python-dotenv
faiss-cpu
sentence-transformers
huggingface_hub
openai
langchain-openai
langchain-huggingface
langchain-community
langchain_core
pydantic
fastapi
uvicorn
requests
# Optional: If you plan to use Docker, include the following line
# docker>=6.1.3

# To set up and run the project, follow these steps:
#python -m venv venv
#.\venv\Scripts\activate
#pip install --upgrade pip
#pip install -r requirements.txt

#python AgentRoot.py

#pip install langchain langchain-community transformers huggingface_hub openai python-dotenv
#pip install torch
#pip install torch --index-url https://download.pytorch.org/whl/cpu
#pip install scikit-learn pandas regex
#For RAG tool
#pip install faiss-cpu
# pip install sentence-transformers
#pip install -U langchain-openai langchain-huggingface langchain-community
#pip install colorama
#pip install rich

